Moved Jackson exception handling into controller advice.
